Add function that generates VMD script of vector field
Can be overlayed with a isosurface if you'd like
TCL function is defined that creates an arrow based on its center,
direction, length, and the radii of the cylinder and cone from which the
arrow is built. The cone is assume to be "equilateral" (intersection of
cone with plane through its center always results in an equilateral
triangle).
As the "weight" of the vector in a vector field changes, the radii and
the length need to change accordingly. This behaviour is goverened by
python function `decompose_weight`.

TODO : `decompose_weight` needs to be fine tuned