Update QA to not check_name merged commits #144
Merged commits have "GitHub <noreply@github.com>" as committer which is
not included in the AUTHORS file of ChemTools. This causes check_name.py
script to fail when running QA on travis for a PR, because the merged
commit of origin/master and user_remote/feature has "GitHub
<noreply@github.com>" as committer.
The --no-merges flag ignores all merged commits when obtaining list of
names.