Move files in orbtools directory to appropriate chemtools module
The orbital partitioning tools were developed in a separate repository
(mainly for compatibility to tox and python36) i.e.
https://github.com/kimt33/chemtools-orbpart
Actually, much of the code was copied from another repository
https://github.com/quantumelephant/dumbo
Nonetheless, it was easier to rebase the chemtools-orbtools repository
onto chemtools than it was to cherry-pick and modify each commit. The
appropriate files are now being moved over to the correct locations