Vectorize Hessian construction in NCI
in chemtools.toolbox.interactions.NCI, the hessian needs to be converted
from the compact, one-dimensional form (only the upper triangular pieces)
to a two-dimensional (square) form from which we obtain the eigenvalues.
The little tricky bit is that this must be repeated for each grid point
(dimension 0).

So we first construct a flatten the hessian part (dimensions 1 and 2) of
the 3-dimensional array, then hard-coded the appropriate parts of the
hessian to this array assuming C-ordering. I think this will fail if the
hessian is not C-ordering