Add indexing when constructing density matrices
This method is needed in other methods that need to use the density matrices but
also need to select specific orbitals.

Note:
The default value for the index, i.e. all orbitals, is different than the ones
for orbitals, i.e. all occupied orbitals. I'm pretty sure this is the correct
default behaviour for the methods that will need to use this functionality