Add function that makes a mesh given three coordinates
This function can be used to make a 2d contour plots of any scalar property.

Issue:
API is kind of awful. I tried to infer the appropriate mesh shape from the given
points along the plane. It is one thing to get the equation for the plane, but
another to find the correct mesh orientation within the plane. It needs to have
the correct center and the size. At the moment, the center is the centroid of
the triangle formed by the three points and the size is given by the distance
away from the center. Maybe it'l be better to define more parameterss to make it
more customizable