Try to fix HORTON log issue (again)
Turns out there is a horton warning that we cannot avoid, which means
that if we set the horton.log to silent, horton will raise RuntimeError.
The next level in log, warning, will still print the banner and the
footer.

Since we cannot fix horton here without being too messy, we'll over
write the header and footer with nothing